semantic search engine (1)/requirements.txt
joblib
pandas
azureml-sdk
numpy
transformers
sentence-transformers
torch